UCP/WIREUP: Split lane selection logic to another file.
